Feature: Multiline Trigger can now take Lua functions as conditions Feature: Multiline Trigger can now take Begin of line substrings as condition Fix: profiles were not ordered correctly according to their date which led to the wrong profile being loaded. Project: started work on enabling autosaving in the trigger editor.

+++ src/TBuffer.cpp
@@ -95,7 +95,7 @@ int TBuffer::getLastLineNumber()
 
 void TBuffer::append( QString text, QColor & fgColor, QColor & bgColor, bool bold, bool italics, bool underline )
 {
-    if( buffer.size() > mLinesLimit )
+    if( (int)buffer.size() > mLinesLimit )
     {
         while( buffer.size() > mLinesLimit-10000 )
         {
